requirements.txt
altgraph==0.17.4
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
exceptiongroup==1.3.0
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
macholib==1.16.3
outcome==1.3.0.post0
packaging==25.0
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.4
selenium==4.33.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
undetected-chromedriver==3.5.5
urllib3==2.4.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
zipp==3.23.0
